Reorg config, fixes tests
